Changed handleAirFrame parameter to AirFrame pointer instead of cMessage. Added getState() method to BaseBattery. PhyBattery checks now the battery state on message handling, if the host state will be set unequal to ACTIVE then the phy goes to sleep state.

+++ src/modules/mobility/CircleMobility.cc
@@ -40,7 +40,7 @@ void CircleMobility::initialize(int stage)
         r = par("r");
         ASSERT(r>0);
         angle = par("startAngle").doubleValue()/180.0*PI;
-        move.setSpeed(par("speed"));
+        move.setSpeed(par("speed").doubleValue());
         omega = move.getSpeed()/r;
 
         // calculate initial position
